Fix paddings and reorganize the code
